Build: minify the JS plugin worker
Add a bogus bootstrap module that configures RequireJS & invokes the actual worker,
to appease r.js optimizer.

Allow modules in build config to specify their "caller" (ie. the HTML or js file
that loads them). This is required for workers, as our former assumption of the
caller being {module}.html does not hold for workers.