Bugs 434865 and 434868: Added templateCollector and templateExplorer.
Defined a new extension point, orion.editor.templates, to simplify the addition of new templates via plugin.

TemplateCollector gathers all the template providers via this extension point and registers itself as a ContentAssistProvider.

TemplateExplorer uses the templateCollector to display the templates in the sidebar and is based on the outline view.

Ported existing templates for javascript, css, and html to make use of the new extension point.  Added new parameter to template objects: 'contentType'.

Change-Id: I35f2d4f642d9ac7d5d5b68c2b189a09e81fb4fbb
Signed-off-by: Ethan Groves <egroves@us.ibm.com>