Bug 477600 - [search] Provide a search option to only see results that match the whole word / regex
3 logical areas needed to be addressed in the client:

(1) UI changes to display "Whole Word" check box and logic to register the response as well as adding "Whole Word" flag to server request.
(2) When clicking on file name in search results the editor needs to highlight correct match.
(3) Expanding matches. For this the RegEx search logic was used with the addition of escaping meta characters for a string literal search. 

In general the whole word matching is using the word boundary assertion "\b" following the convention of the functionality that already exists in Edit->Find under Options as "Whole Word".

Change-Id: I85bb68b0f671f82bc2ef8d7b013b9ae7108f364d
Signed-off-by: Dmitry Shindyalov <rushingrussian@yahoo.com>