Bug 507797 - Viewing the compared content of an untracked file throws 404
When looking at changes before staging or committing, the Git
integration will render a diff between the index and the working
directory. If a file is untracked, it will not exist in the index. If
we try to query for the content of the file in the index to run the
diff, nodegit won't find anything because an untracked file does not
exist in the index. We should consider this case and return "" for the
file's content in the index as it does not exist there

Change-Id: Idcc6ba64c20c5b44daef2518b352eea2e75fa38c
Signed-off-by: Remy Suen <remy.suen@gmail.com>