Bug 507797 - Viewing the compared content of an untracked file throws 404
When asking to look at a file in the Git index, read the
'read-if-exists' property in the XHR to determine whether it is
acceptable or not for the file to not actually exist. If the property
is 'true' and the file doesn't exist, respond back with a 204. In all
other cases where the file doesn't exist, send a 404 response back to
the client.

Change-Id: If8b082814236c89587f7cd2ac0166d4f1465ba06
Signed-off-by: Remy Suen <remy.suen@gmail.com>