Bug 507934 - Tasks can be deleted by anyone
Record the user that created the task in the task itself. When a client
asks for a task to be deleted, check that the user id of the client
requesting client asking for the task to be deleted shares the same
user id as the task's originator. If it's a match, delete it. If not,
return a 404 to the client.

Change-Id: I39eac0629bf574004166f1b480373c070319d8e5
Signed-off-by: Remy Suen <remy.suen@gmail.com>