Bug 507934 - Prevent access of other user's tasks
Currently, any user can retrieve or delete another user's task as long
as the task's id is known. To prevent this, we record the user that
initiated the task in the task itself. Then, when a client asks for a
task, check that the user id of the client requesting the task shares
the same user id as the task's originator. If it's a match, perform the
requested operation. If not, return a 404 to the client.

Change-Id: I39eac0629bf574004166f1b480373c070319d8e5
Signed-off-by: Remy Suen <remy.suen@gmail.com>