Bug 507334 - 'git log' spins forever in Node/Electron
The previous implementation of 'git log' was extremely inefficient in
handling files. It would fetch every single commit in full and then
check the diff to determine whether it affected the interested file.

The new approach is to first iterate over all the commits quickly to
get a count of how many commits are in the current branch. Once that
has been recorded, use fileHistoryWalk(*) over the entire history to
grab all the commits that affect the interested file. From there, we
further trim it down to the paging size and only fetch the full commits
of what we need for the JSON response.

Change-Id: I104f98c51e25b8f31ce2578fb1ecebe054bcf615
Signed-off-by: Remy Suen <remy.suen@gmail.com>