Bug 508511 - 'git log' throws many exceptions that are ignored when looking for refs and tags
When responding to a log request, we want to let the client know about
any refs that a given commit is pointing to. We do this in
getCommitRefs(*) by fetching all the references in the repository and
checking if the commit that ref is pointing to is one of the commits
that we are sending back to the client.

The current code catches and ignores all errors when processing the
refs. Since there are likely many refs in a repository that is pointing
to commits both old and new, it is highly unlikely that all of them
will be pointing at one of the commits to be included in the current
response. We should check whether a given ref is pointing at a commit
that is being sent back to the client before processing it instead of
blatantly trying to process every single commit that a ref is pointing
at.

Change-Id: I8cbdd2f81f880395b0311d3a06306f501630dabb
Signed-off-by: Remy Suen <remy.suen@gmail.com>