Bug 508584 - Yellow 'ALREADY_UPDATE' status message when merging nothing new to the active branch
When a merge happens, the Orion client is expecting an
"ALREADY_UP_TO_DATE" string to be returned to indicate that no merging
was required. Both are constants that have been defined by JGit's
MergeResult.MergeStatus enum. Fix the Node server to return this same
string if such cases have been detected after a merge.

Change-Id: I2b11a727e03fe7ce9f3c26b7ad362ea20341592b
Signed-off-by: Remy Suen <remy.suen@gmail.com>