Bug 510950 - Failed merges do not put the repository in a merge state on Node
When a merge fails via Repository's mergeBranches(*) function from
NodeGit, the operation is abandoned completely. The repository will not
be marked as being in a merge. Instead, a conflicted index is provided
for inspection to the API consumer.

To correct this, we should call Merge's merge(*) function from NodeGit
when we realize that the merge has failed due to conflicts. This will
bring in the desired commit into HEAD with any conflicts marked as such
within the working directory (instead of simply providing an in-memory
index with conflicts) in addition to putting the repository into a
merge state by creating the relevant .git/MERGE_* files.

Signed-off-by: Remy Suen <remy.suen@gmail.com>