Bug 511295 - Merge commits cannot be created on Node
When a commit is created, the code currently always builds the commit
on top of the HEAD commit (except for the amend case obviously).
However, if the repository is in a merge state, we should instead be
iterating over the MERGE_HEAD file and bringing those parents in as
well so that the created commit is parented off of the HEAD commit
and the commits that have been recorded in the MERGE_HEAD file. Once
the merge commit has been created successfully, the MERGE_* files
within the .git/ folder should then be deleted as they are no longer
needed.

Signed-off-by: Remy Suen <remy.suen@gmail.com>