Bug 511076 - Checking a file to commit while unresolved conflicts exist anywhere causes an exception
When a PUT request comes in for the index, the index object must be
written onto disk. However, it is not necessary to write the index as a
tree representation as it may or may not yet be ready to be committed.
The call to writeTree() can be performed when the client actually sends
a POST request to the commit. Preemptively writing the index as a tree
for any PUT request may fail as the user might not have resolved all
the conflicts in the index.

Signed-off-by: Remy Suen <remy.suen@gmail.com>