Bug 511824 - All tags are flagged as being lightweight
When the client requests for the history of a repository or a file, a
list of commits will be returned to the user. The returned JSON
includes information about any tags associated with a given commit.
However, those are all being incorrectly flagged as lightweight tags.
Changed the code to identify whether a tag is annotated or not and
to flag it as such in the returned JSON to the client.

Signed-off-by: Remy Suen <remy.suen@gmail.com>