Bug 511822 - Upgrade NodeGit to 0.17.0
On NodeGit 0.17.0, comparing two branches will cause Node to crash
due to issue #4102 in libgit2. This is an essential use case of
Orion's Git integration as users need to be able to see the incoming
and outgoing changes between the different branches. This regression
test should hopefully catch this error should this bug surface again
in another form in the future.

Signed-off-by: Remy Suen <remy.suen@gmail.com>