Bug 513503 - Can't checkout local branch that has the same name as a tag
If a branch and a tag shares the same name, Git prioritizes the
resolution of a tag before a branch. This causes problem if a user is
trying to perform a checkout or any other branch-related operations.
To get around this, we will specifically append `refs/heads` to the
name of the branch that the client has specified to ensure that Git
will resolve the reference to the branch instead of the tag.

Signed-off-by: Remy Suen <remy.suen@gmail.com>