Bug 513537 - Checking out branch with an encoded name leaves Git page inaccessible
When information is transferred from the server to the client, URLs
contained in the JSON are escaped two times. When the client uses
these URLs to send a request to the server, the server will decode it
once so the URL is still in a singly encoded state. The server only
needs to perform one decoding pass after that instead of two.

Signed-off-by: Remy Suen <remy.suen@gmail.com>