Debugger implementation
This commit contains both the frontend and the backend.

The frontend contains a debug service, a slideout pane in the side bar
(debug UI) and a deploy service.

The backend talks to debug adapters which implement VSCode Debug
Protocol. it can be used either as a module of Orionode or a standalone
server.

Change-Id: Ib6f17715f498301fa699bfad86a49ca6e667c7c9
Signed-off-by: Martin Wang <jiahaow@ca.ibm.com>