Bug 514357 - `git rebase --skip` is unimplemented
When a rebase is performed, changes in the active branch may conflict
with the other branch. In some cases, it may be desirable to skip and
discard a change in the active branch.

When a skip happens, any untracked files shall be deleted and any
changes in the working directory or index will be discarded. After
that has been completed, the current commit will be skipped over and
rebasing will resume starting from the next commit to rebase. If
there are no other commits left to rebase, the rebasing will be
marked as being completed.

Signed-off-by: Remy Suen <remy.suen@gmail.com>