Bug 510062 - Git log graph is broken on Node
If the list of commits that affect a given file exceed the page size,
the parent information of some commits may be incorrect because they
are pointing at the commit's actual parent and not a parent commit
that affects the interested file.

In such a situation, we first iterate and find all the parent commits
that are pointing at something that isn't being included in the
returned JSON response. For each one of those parents, we walk
through the tree iterating over their parents until we find a parent
that actually affects the given file. If such a commit has been
found, we then update the JSON information so that it now points to
this other parent commit that affects the interested file.

Signed-off-by: Remy Suen <remy.suen@gmail.com>