Bug 510062 - Git log graph is broken on Node
A RevWalk object appears to remember some history information about
its prior runs when reused even if a new commit has been pushed onto
it as a starting point. To prevent this from happening, a new RevWalk
will be instantiated for each unresolved parent commit to prevent
commits from other branches to be included when iterating back in
time.

Signed-off-by: Remy Suen <remy.suen@gmail.com>