Bug 515428 - Can't calculate incoming/outgoing changes between two unrelated histories
To determine the incoming and outgoing changes between two branches,
a common ancestor is used for determining when the two branches
diverged and which commits they don't share. In the event of two
unrelated histories, the two branches will not have a common ancestor
commit. In such a case, all the commits of each individual branch
should be considered as being incoming/outgoing.

Signed-off-by: Remy Suen <remy.suen@gmail.com>