Bug 510062 - Git log graph is broken on Node
When considering merge commits that affect a given file, a given
branch may be ignored if none of its changes matter for the given
file. In such a case, the common ancestor is used to record when the
code should stop ignoring commits. However, if a merge commit is
between two different unrelated histories, there will be no common
ancestors. In such a case, don't bother flagging the (null) common
ancestor as something to be recorded as the entire branch shall just
be ignored outright.

Signed-off-by: Remy Suen <remy.suen@gmail.com>