Bug 510646 - Completed operations don't actually get removed by operationsClient
When the operations client tries to remove completed operations, it
retrieves the current collection of operations, removes the completed
ones, and sets the modified collection back. However, this operation
doesn't actually work because the preferences services assumes that a
modification is being made to the passed in collection and that the
entries that have been removed from the collection should not be
acted upon. To fix this, the clear flag will be defined and set to
true so that the preferences service will know that the preferences
should be overwritten with the passed in collection.

Signed-off-by: Remy Suen <remy.suen@gmail.com>