Bug 517230 - Provide content assist for the escape directive in Dockerfiles
If the user's text caret is currently inside the first comment of the
Dockerfile and there are no instructions preceding this first
comment, we should suggest the escape directive as a content assist
proposal if the user invokes content assist.

The user should also be provided with documentation assistance if the
user hovers over a comment that looks like an escape directive.

Signed-off-by: Remy Suen <remy.suen@gmail.com>