Bug 517525 - Dockerfile content assist parser should ignore invalid escape parser directive declarations
If the Dockerfile currently being edited defines an escape parser
directive that is not a backslash or a backtick or if it is not
defined correctly, the content assist parser should simply use the
original backslash character as the escape character.

This fallback behaviour brings the content assist parser in line with
the behaviour of the Dockerfile builder where an escape directive
that is not a backtick or a backslash will cause the builder to fail
and not build the Docker image. If on the other hand the escape
parser directive is properly declared but nothing follows the
declaration except a newline character, such declarations too will
also be ignored and the backslash character will be used as the
newline character instead.

Signed-off-by: Remy Suen <remy.suen@gmail.com>