Bug 517621 - Stash REST API gives ambiguous errors if the stash is empty
If the stash is empty and a client tries to contact the Orion server
to drop the entire stash or a given stash revision, the server should
send back an error message that indicates that the stash is empty.

If the stash is not empty and a client tries to drop the stash with
an invalid stash revision, the server should send back an error
message to notify the user that the provided stash revision is
invalid and that the specified stash revision could not be found in
the repository's list of refs.

Signed-off-by: Remy Suen <remy.suen@gmail.com>