Bug 517690 - Stash REST API gives ambiguous errors when handling PUT requests
If a client tries to apply something from the stash when the stash is
actually empty, the client should be told that nothing could be
applied because the stash is empty.

Similarly, if a client tries to apply a stash revision that doesn't
actually exist in the stash, the client should be told that the
provided stash revision is invalid.

Signed-off-by: Remy Suen <remy.suen@gmail.com>