Prevent the LSP registry from trying to read unknown content types
As a user can open any sort of file with any file extension in Orion,
it is impossible to guarantee that every file that is opened will
have a known content type. The language server registry should
not automatically assume that a passed in content type is non-null and
valid when trying to retrieve a language server for a given content
type.

Signed-off-by: Remy Suen <remy.suen@gmail.com>