Bug 518438 - [lsp] Content assist won't work when trying to use Java's '.' trigger character
If a ContentAssistProvider instance is wrapping around a language
server, we should not tamper with the lspServer flag by deleting it.
Removing the flag causes the content assist code to no longer
recognize that the current provider is wrapping around a language
server and as a result will fail to compute proposals when using
trigger characters to initiate content assist.

Signed-off-by: Remy Suen <remy.suen@gmail.com>