Bug 518631 - [lsp] Orion does not consider TextDocumentSyncOptions from the server
When a document is modified by Orion, it should send
'textDocument/didChange' notifications to the server so that it knows
that the document has changed. However, depending on the server's
setting, Orion should choose to either a) not send any notifications
at all if TextDocumentSyncKind.None has been specified or b) send the
document in its entirety if the server only supports
TextDocumentSyncKind.Full.

Signed-off-by: Remy Suen <remy.suen@gmail.com>