Bug 518679 - [lsp] Outline ignores the location of SymbolInformations if containerName not specified
As the containerName property of a SymbolInformation is optional,
language servers are not obligated to include it when sending a
response back to the client. As such, Orion should not assume that
the containerName property has been specified and should make an
effort to present an outline that is as functional as possible to the
user even when no containerName information is present.

Signed-off-by: Remy Suen <remy.suen@gmail.com>