Fix: Collab Highlight width inconsistency
Collab highlight width was inconsistent across different browsers.
The width would vary on different browsers as peers select long texts
on the same line. This is due to the use of a calculated constant in the code
rather than using environment variables that we were unaware of at the time.
Thanks to Silenio, we found a variable to fixes this issue.