Bug 527206 - Merge LSP support into master - contentAssist.changes
This does not include the necessary changes to make the LSP's
CompletionItem's additionalTextEdits work.

The current code in utils.js uses a property named additionalEdits
which conflicts with a property that was created for people extending
Orion to add additional text edits to a completion proposal.

See bug 439272 and bug 527206 comment 14 for more details.

Signed-off-by: Remy Suen <remy.suen@gmail.com>