Bug 527206 - Merge LSP support into master
The proposals from a language server were all being set with an
escapePosition field. These were always defined as being the length
of the proposal's text and the code would later perform offset
calculations based on the CompletionItem's textEdit object. However,
if a CompletionItem did not have a textEdit defined, the escape
position would be completely off as it would be relative from zero
instead of from the content that was being replaced by the proposal.

The fix is to remove the property from proposals created from
CompletionItems that don't have a textEdit. This will allow Orion to
calculate and adjust itself to the escape position based on the
replaced text.

Signed-off-by: Remy Suen <remy.suen@gmail.com>