Bug 531065 - Errors in server unit tests on Travis / Linux (#300)
* Bug 531065 - Errors in server unit tests on Travis / Linux

Update tests to fail correctly when running on Linux. The tests in
question are supposed to fail based on max path length, updated test
paths to exceed 4096 characters (limit in nix systems).

Signed-off-by: Casey Flynn <caseyflynn@google.com>

* Bug 531065 - Errors in server unit tests on Travis / Linux

Update tests to fail correctly when running on Linux. The tests in
question are supposed to fail based on max path length, updated test
paths to exceed 4096 characters (limit in nix systems).

Signed-off-by: Casey Flynn <caseyflynn@google.com>

* Add option to grunt task mocha_istanbul.

Currently the mocha_istanbul task does not exit as expected causing
Travis to break.

Signed-off-by: Casey Flynn <caseyflynn@google.com>